refactor one-off out of utils.py
